Library structure refactored
