+++ src/methods/bayes/base/output.py
