+++ src/methods/bayes/variational/gaussian_variational.py
