+++ src/methods/bayes/variational/optimization_gaussian.py
