+++ src/methods/report/base.py
